Flesh out options, steps.
