--- app/components/attribute-list/attribute-list.pug
@@ -1,1 +1,0 @@
-section(data-component="attribute-list" data-flow="column")
