--- app/components/attribute-list/attribute-list.styl
@@ -1,41 +1,0 @@
-@import "./styl/colors"
-@import "./styl/fonts"
-@import "./styl/font-smoothing"
-
-[data-component="attribute-list"]
-
-  > p:first-child
-    padding-top 2em
-
-  .form
-    flex 1 1 auto
-
-    .standard
-      flex 1 0 auto
-      width 100%
-
-  .entity
-    padding 1.25em
-
-    + .entity
-      border-top 1px solid #eee
-
-  .details
-    width 27%
-    margin-right 5%
-    padding-top 0.15em
-
-    .identifier, .token
-      antialiasedFonts()
-      font-family monospaceFonts
-      line-height 1.6
-
-    .token
-      color subtleText
-      display inline-block
-      overflow hidden
-      text-overflow ellipsis
-      white-space nowrap
-
-  .label
-    display block
