+++ app/components/steps/creating/creating.pug
@@ -1,2 +1,9 @@
 section.step(data-step="creating" data-flow="column")
-  p creating
+  header.header(data-flow="column")
+    h1 Packaging your plugin...
+
+    .loading-dots(data-state="loading")
+      i
+      i
+      i
+
