+++ app/components/steps/creating/creating.styl
@@ -1,2 +1,12 @@
 .step[data-step="creating"]
-  color red
+  .header
+    align-items center
+    text-align center
+
+    h1
+      margin-top 2em
+      font-weight normal
+
+    .loading-dots
+      margin-left -1.5em
+
