+++ app/components/steps/download/download.styl
@@ -1,2 +1,18 @@
 .step[data-step="download"]
-  color red
+
+  .header
+    align-items center
+    text-align center
+
+    h1
+      margin-top 2em
+      font-weight normal
+
+    .download-link
+      text-align center
+      font-size .9em
+      opacity 0.75
+
+  .content
+    width 50%
+    margin 1em auto
