+++ app/components/steps/embed-code/attribute-picker/attribute-picker.styl
@@ -15,6 +15,9 @@
   // Fix animation artifacts.
   subpixelAntialiasedFonts()
 
+  span
+    subpixelAntialiasedFonts()
+
   [data-entity-id]
     display inline-block
     position relative
