+++ app/components/steps/embed-code/embed-code.pug
@@ -1,4 +1,4 @@
-section.step(data-step="embedCode" data-flow="column")
+section.step.padded(data-step="embedCode" data-flow="column")
   header.column-details
     .column(data-flow="column")
       h1.headline
