+++ app/components/steps/embed-code/embed-code.styl
@@ -2,8 +2,6 @@
 @import "~styl/font-smoothing"
 
 .step[data-step="embedCode"]
-  padding 0 2em
-
   .column
     flex 1 1 auto
     width 50%
