+++ app/components/steps/schema/entity.pug
@@ -1,4 +1,4 @@
-.entity
+.entity.box
   .details(data-flow="column")
     code.identifier= entity.identifier
     code.token= entity.original
