+++ app/components/steps/schema/schema.pug
@@ -1,2 +1,10 @@
-section.step(data-step="schema" data-flow="column")
-  p foo
+section.step.padded(data-step="schema" data-flow="column")
+  header.step-details(data-flow="column")
+    h1.headline
+      .step-number 2
+      .label Customize these dynamic options.
+
+    p
+      | This is the embed code your users add to their website to install your tool or service.
+
+  .property-list(data-ref="propertyList" data-flow="column")
