+++ app/components/steps/schema/schema.styl
@@ -1,2 +1,42 @@
+@import "./styl/colors"
+@import "./styl/fonts"
+@import "./styl/font-smoothing"
 .step[data-step="schema"]
-  color blue
+  .step-details
+    flex 0 0 auto
+
+  .property-list
+    flex 0 0 auto
+
+  .form
+    flex 1 1 auto
+
+    .standard
+      flex 1 0 auto
+      width 100%
+
+  .entity
+    padding 1.25em
+
+    + .entity
+      margin-top 1em
+
+  .details
+    width 27%
+    margin-right 5%
+    padding-top 0.15em
+
+    .identifier, .token
+      antialiasedFonts()
+      font-family monospaceFonts
+      line-height 1.6
+
+    .token
+      color subtleText
+      display inline-block
+      overflow hidden
+      text-overflow ellipsis
+      white-space nowrap
+
+  .label
+    display block
