+++ app/styl/colors.styl
@@ -5,7 +5,6 @@ darkOffwhite = #fafafa
 
 offBlack = #3b3b3b
 
-// darkGray = #2b2b2b
 darkGray = #616161
 gray = #e0e0e0
 lightGray = #ccc
